mysite/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
html2markdown==0.1.7
idna==3.10
Markdown==3.7
markdownify==0.13.1
mysqlclient==2.2.4
pillow==11.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
